Add a callTrace level and is*Enabled() functions
The callTrace level is lower than debug, not activated by default
and could be used by a call tracer.

New function is*Enabled() (where * is a level in title case) could
be used to check if a given level is currently enabled, and allow
caller to skip computation only needed for logging when logging is
disabled.